GameObject.GetComponent
Added static method to Game Object class to allow other objects to get a component and interact with it. This in combination with makink the Input Response on the Game Controller a static field will eliminate the need for passing the game controller to the action methods. Instead we will pass the ParseTree object which contains the verb the direct object and the indirect object if applicable.